+++ processor/src/main/java/react4j/processor/ReactProcessor.java
@@ -1261,13 +1261,14 @@ public final class ReactProcessor
             parameters.contains( "priority" ) &&
             parameters.contains( "keepAlive" ) &&
             parameters.contains( "observeLowerPriorityDependencies" ) &&
-            4 == parameters.size() ) )
+            parameters.contains( "arezOnlyDependencies" ) &&
+            5 == parameters.size() ) )
     {
       throw new ReactProcessorException( "The @" + Constants.COMPUTED_ANNOTATION_CLASSNAME + " annotation was " +
-                                         "expected to have the parameters name, priority, keepAlive and " +
-                                         "observeLowerPriorityDependencies but has " + parameters + ". The " +
-                                         "react4j annotation processor needs to be updated to handle " +
-                                         "the change in parameters.", typeElement );
+                                         "expected to have the parameters name, priority, keepAlive, " +
+                                         "arezOnlyDependencies and observeLowerPriorityDependencies but has " +
+                                         parameters + ". The react4j annotation processor needs to be updated to " +
+                                         "handle the change in parameters.", typeElement );
 
     }
   }