+++ processor/src/test/resources/expected/com/example/arez/computed/React4j_ParameterizedComputedComponent.java
@@ -42,6 +42,7 @@ abstract class React4j_ParameterizedComputedComponent extends ParameterizedCompu
       priority = Priority.LOWEST,
       name = "foo",
       keepAlive = true,
+      arezOnlyDependencies = false,
       observeLowerPriorityDependencies = true
   )
   boolean isActive() {